migrate `send_password_reset_mail` to `.mail` and delete `_old`
